python3.12/django/app_23/pyproject.toml
[tool.poetry]
name = "app-23"
version = "0.1.0"
description = ""
authors = ["noskill <golanghack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
kafka-python = "^2.0.2"
djangorestframework-simplejwt = "^5.3.1"
swagger-ui = "^0.1.2"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
